fix: move custom fee logic
