+++ src/common/transactions/transactions.ts
@@ -11,6 +11,15 @@ import {
   ContractCallPayload,
   ContractDeployPayload,
 } from '@stacks/connect/dist/types/types/transactions';
+import BN from 'bn.js';
+import { isNumber, isString } from '@common/utils';
+
+function safelyExtractFeeValue(fee: unknown) {
+  if (fee === '') return undefined;
+  if (!isNumber(fee) && !isString(fee)) return undefined;
+  if (!Number.isFinite(parseInt(String(fee)))) return undefined;
+  return new BN(fee, 10);
+}
 
 interface BroadcastTransactionOptions {
   txRaw: string;
@@ -69,7 +78,7 @@ export type ContractCallOptions = Pick<
   | 'postConditionMode'
   | 'postConditions'
   | 'network'
->;
+> & { fee?: number | string | BN };
 
 export interface TokenTransferOptions {
   txType: TransactionTypes.STXTransfer;
@@ -77,12 +86,13 @@ export interface TokenTransferOptions {
   memo: STXTransferPayload['memo'];
   amount: STXTransferPayload['amount'];
   network: STXTransferPayload['network'];
+  fee?: number | string | BN;
 }
 
 export type ContractDeployOptions = Pick<
   ContractDeployPayload,
   'txType' | 'contractName' | 'codeBody' | 'postConditions' | 'postConditionMode' | 'network'
->;
+> & { fee?: number | string | BN };
 
 interface GenerateSignedTransactionOptions {
   txData: ContractCallOptions | TokenTransferOptions | ContractDeployOptions;
@@ -92,10 +102,14 @@ interface GenerateSignedTransactionOptions {
 }
 
 export async function generateSignedTransaction(options: GenerateSignedTransactionOptions) {
-  const { txData, senderKey, nonce, fee } = options;
+  const { txData, senderKey, nonce } = options;
   const isValid = getIsValid(txData.txType);
 
   if (!isValid) throw new Error(`Invalid Transaction Type: ${txData.txType}`);
+
+  // TODO: how to overwrite
+  const customAppFee = safelyExtractFeeValue(txData.fee);
+  const fee = customAppFee?.toNumber() || options.fee;
 
   switch (txData.txType) {
     case TransactionTypes.ContractCall:
